A script to tell the time via festival ? text to speach system. Initial release.
This will check that the user is logged in on the console before
trying to send the text to be played over the speakers.

You will have to have  ?? something installed and configured correctly
before this will work.

This should be run (by command line or from cron) by the user that
will login into X on the console.

Have fun.

==>brian.