A script to tell the time via the speakers.
This script will tell the time via the speakers using speachd.  The
script to check to make sure the user is logged on the console
in X before trying to tell the time via speechd.

You need to be logged in so the framework of the audio device is
setup to play the audio.  You also don't want to try and play something
if you aren't logged in because this could mess up your audio (it
may take control of the audio device so the sound system ?pulseaudio?
will not be able to use it when you do login.. arg.) so audio
wouldn't work when you do login.

Inital release.

Good Luck, have fun.

==>brian.